refactor: Remove @simple-bookkeeping/types package and migrate type definitions (#395)
- Removed entire packages/types directory
- Migrated essential types to apps/web/src/types/:
  - SimpleEntry types for demo feature
  - Journal DTOs for Server Actions
  - API error types
- Inlined enum definitions in packages that need them:
  - AccountType, JournalStatus in shared package
  - User, Account types in test-utils
- Updated all import paths to use new locations
- Removed @simple-bookkeeping/types dependencies from all package.json files
- Fixed tsconfig.json references to removed package

This simplifies the codebase by eliminating an unnecessary package and
leveraging TypeScript's type inference capabilities with Server Actions.
