+++ apps/web/package.json
@@ -43,7 +43,6 @@
     "@simple-bookkeeping/shared": "workspace:*",
     "@simple-bookkeeping/supabase-client": "workspace:*",
     "@simple-bookkeeping/test-utils": "workspace:*",
-    "@simple-bookkeeping/types": "workspace:*",
     "@supabase/auth-helpers-nextjs": "^0.10.0",
     "@supabase/auth-helpers-react": "^0.5.0",
     "@supabase/ssr": "^0.5.1",
