+++ apps/web/src/types/api.ts
@@ -1,15 +1,6 @@
 /**
  * API-specific type definitions
- * Note: Express types have been removed. Use Next.js request types instead.
  */
-
-// import { User } from './auth';
-
-// This interface has been removed as it depends on Express.
-// For Next.js, use NextRequest or the request object from API routes.
-// export interface AuthenticatedRequest extends Request {
-//   user?: User;
-// }
 
 export interface ValidationError {
   field: string;
@@ -38,6 +29,3 @@ export interface ApiServerError {
   code: 'INTERNAL_SERVER_ERROR';
   message: string;
 }
-
-// Export API subdirectory types
-export * from './api/journal-entries';
