+++ apps/web/src/types/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Type definitions for the web application
+ * These types have been migrated from the deprecated @simple-bookkeeping/types package
+ */
+
+export * from './simple-entry';
+export * from './journal';
+export * from './api';
