+++ apps/web/src/types/simple-entry.ts
