+++ packages/test-utils/package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@faker-js/faker": "^8.4.1",
-    "@simple-bookkeeping/types": "workspace:*",
     "jsonwebtoken": "^9.0.2"
   },
   "devDependencies": {
