+++ packages/test-utils/tsconfig.json
@@ -7,6 +7,5 @@
     "declaration": true,
     "declarationMap": true
   },
-  "include": ["src/**/*"],
-  "references": [{ "path": "../types" }]
+  "include": ["src/**/*"]
 }
