--- packages/types/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@simple-bookkeeping/types",
-  "version": "1.0.0",
-  "description": "Shared TypeScript types for Simple Bookkeeping",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch"
-  },
-  "dependencies": {
-    "zod": "^4.0.14"
-  },
-  "devDependencies": {
-    "typescript": "^5.8.3"
-  }
-}
