--- packages/types/src/api/index.ts
@@ -1,1 +1,0 @@
-export * from './journal-entries';
