--- packages/types/src/common.ts
@@ -1,42 +1,0 @@
-/**
- * Common type definitions
- */
-
-export interface PaginationQuery {
-  page?: number;
-  limit?: number;
-  sort?: string;
-  order?: 'asc' | 'desc';
-}
-
-export interface PaginatedResponse<T> {
-  data: T[];
-  meta: {
-    page: number;
-    limit: number;
-    total: number;
-    totalPages: number;
-  };
-}
-
-export interface ApiResponse<T = unknown> {
-  data?: T;
-  error?: ApiError;
-  meta?: Record<string, unknown>;
-}
-
-export interface ApiError {
-  code: string;
-  message: string;
-  details?: unknown;
-}
-
-export interface DateRange {
-  startDate: string;
-  endDate: string;
-}
-
-export interface MoneyAmount {
-  amount: number;
-  currency?: string;
-}
