--- packages/types/src/index.ts
@@ -1,12 +1,0 @@
-export * from './auth';
-export * from './organization';
-export * from './account';
-export * from './accountingPeriod';
-export * from './journal';
-export * from './partner';
-export * from './ledger';
-export * from './report';
-export * from './api';
-export * from './common';
-export * from './enums';
-export * from './simple-entry';
